+++ src/sigmaepsilon/core/misc.py
